Разбил код на ES-модули
